# cookies: improve validateCookieName