--- a/benchmarks/cookies/validate-cookie-name.mjs
+++ b/benchmarks/cookies/validate-cookie-name.mjs
@@ -1,0 +1,12 @@
+import { bench, group, run } from 'mitata'
+import { validateCookieName } from '../../lib/web/cookies/util.js'
+
+const valid = 'Cat'
+
+group('validateCookieName', () => {
+  bench(`valid: ${valid}`, () => {
+    return validateCookieName(valid)
+  })
+})
+
+await run()