--- a/lib/web/cookies/util.js
+++ b/lib/web/cookies/util.js
@@ -32,28 +32,29 @@ function isCTLExcludingHtab (value) {
  * @param {string} name
  */
 function validateCookieName (name) {
-  for (const char of name) {
-    const code = char.charCodeAt(0)
+  for (let i = 0; i < name.length; ++i) {
+    const code = name.charCodeAt(i)
 
     if (
-      (code <= 0x20 || code > 0x7F) ||
-      char === '(' ||
-      char === ')' ||
-      char === '>' ||
-      char === '<' ||
-      char === '@' ||
-      char === ',' ||
-      char === ';' ||
-      char === ':' ||
-      char === '\\' ||
-      char === '"' ||
-      char === '/' ||
-      char === '[' ||
-      char === ']' ||
-      char === '?' ||
-      char === '=' ||
-      char === '{' ||
-      char === '}'
+      code < 0x21 || // exclude CTLs (0-31), SP and HT
+      code > 0x7E || // exclude non-ascii and DEL
+      code === 0x22 || // "
+      code === 0x28 || // (
+      code === 0x29 || // )
+      code === 0x3C || // <
+      code === 0x3E || // >
+      code === 0x40 || // @
+      code === 0x2C || // ,
+      code === 0x3B || // ;
+      code === 0x3A || // :
+      code === 0x5C || // \
+      code === 0x2F || // /
+      code === 0x5B || // [
+      code === 0x5D || // ]
+      code === 0x3F || // ?
+      code === 0x3D || // =
+      code === 0x7B || // {
+      code === 0x7D // }
     ) {
       throw new Error('Invalid cookie name')
     }
@@ -285,6 +286,7 @@ function getHeadersList (headers) {
 
 module.exports = {
   isCTLExcludingHtab,
+  validateCookieName,
   validateCookiePath,
   toIMFDate,
   stringify,