- improve performance
- write tests

before:

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/undici$ node benchmarks/cookies/validate-cookie-name.mjs 
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.6.2 (x64-linux)

benchmark       time (avg)             (min … max)       p75       p99      p999
-------------------------------------------------- -----------------------------
• validateCookieName
-------------------------------------------------- -----------------------------
valid: Cat   29.24 ns/iter     (25.81 ns … 854 ns)  26.67 ns  82.46 ns    107 ns

summary for validateCookieName
  valid: Cat
```

after:

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/undici$ node benchmarks/cookies/validate-cookie-name.mjs 
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.6.2 (x64-linux)

benchmark       time (avg)             (min … max)       p75       p99      p999
-------------------------------------------------- -----------------------------
• validateCookieName
-------------------------------------------------- -----------------------------
valid: Cat    9.95 ns/iter       (8.8 ns … 305 ns)   9.96 ns  13.27 ns  18.21 ns

summary for validateCookieName
  valid: Cat
```